python/requirements.txt
paho-mqtt>=2.0.0
numpy>=1.26.0
pandas>=2.2.0
bleak>=0.22.0
matplotlib>=3.8.0
PyYAML>=6.0.0,<7.0.0
scipy>=1.13.0
